backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv

langchain
langchain-core
langchain-community
langchain-groq
groq

requests
aiohttp

google-api-python-client
google-auth
google-auth-oauthlib
